Use Inheritance Contact class;
